[linux] peripherals: move implementation to platform folder.
