+++ cmake/treedata/android/subdirs.txt
@@ -1,16 +1,16 @@
 xbmc/cores/RetroPlayer/process/android  cores/RetroPlayer/process/android
-xbmc/platform/linux                     platform/linux
 xbmc/input/touch                        input/touch
 xbmc/input/touch/generic                input/touch/generic
-xbmc/peripherals/bus/linux              peripherals/bus/linux
 xbmc/peripherals/bus/android            peripherals/bus/android
 xbmc/powermanagement/android            powermanagement/android
 xbmc/filesystem/posix                   filesystem/posix
-xbmc/utils/posix                        utils_posix
+xbmc/utils/posix                        utils/posix
 xbmc/windowing/android                  windowing/android
 xbmc/platform/posix                     posix
+xbmc/platform/linux                     platform/linux
 xbmc/platform/linux/network             platform/linux/network
-xbmc/platform/android/activity          android_activity
-xbmc/platform/android/bionic_supplement android_bionicsupplement
+xbmc/platform/linux/peripherals         platform/linux/peripherals
+xbmc/platform/android/activity          android/activity
+xbmc/platform/android/bionic_supplement android/bionicsupplement
 xbmc/platform/android/network           platform/android/network
 xbmc/platform/android/storage           platform/android/storage
