+++ cmake/treedata/linux/subdirs.txt
@@ -1,14 +1,14 @@
-xbmc/platform/linux        platform/linux
-xbmc/platform/linux/input  platform/linux/input
-xbmc/platform/linux/network platform/linux/network
-xbmc/platform/linux/storage platform/linux/storage
-xbmc/input/touch           input/touch
-xbmc/input/touch/generic   input/touch/generic
-xbmc/peripherals/bus/linux peripherals/bus/linux
-xbmc/powermanagement/linux powermanagement/linux
-xbmc/filesystem/posix      filesystem/posix
-xbmc/utils/posix           utils_posix
-xbmc/platform/posix        posix
+xbmc/platform/posix                 platform/posix
+xbmc/platform/linux                 platform/linux
+xbmc/platform/linux/input           platform/linux/input
+xbmc/platform/linux/network         platform/linux/network
+xbmc/platform/linux/peripherals     platform/linux/peripherals
+xbmc/platform/linux/storage         platform/linux/storage
+xbmc/input/touch                    input/touch
+xbmc/input/touch/generic            input/touch/generic
+xbmc/powermanagement/linux          powermanagement/linux
+xbmc/filesystem/posix               filesystem/posix
+xbmc/utils/posix                    utils/posix
 xbmc/cores/RetroPlayer/process/rbpi cores/RetroPlayer/process/rbpi
 xbmc/cores/VideoPlayer/Process/rbpi cores/VideoPlayer/Process/rbpi
-xbmc/windowing/linux       windowing/linux
+xbmc/windowing/linux                windowing/linux
