+++ xbmc/platform/linux/peripherals/PeripheralBusUSBLibUSB.cpp
