+++ xbmc/platform/linux/peripherals/PeripheralBusUSBLibUSB.h
