+++ xbmc/platform/linux/peripherals/PeripheralBusUSBLibUdev.cpp
