+++ xbmc/platform/linux/peripherals/PeripheralBusUSBLibUdev.h
